ai_quizzer/quiz/serializers.py
# quiz/serializers.py
from rest_framework import serializers
from .models import Quiz, Submission
from django.contrib.auth.models import User

class UserSerializer(serializers.ModelSerializer):
    # Ensure password is write-only
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ['username', 'password']

    def create(self, validated_data):
        # Create a new user with the validated data
        user = User.objects.create_user(
            username=validated_data['username'],
            password=validated_data['password']
        )
        return user

class QuizSerializer(serializers.ModelSerializer):
    class Meta:
        model = Quiz
        fields = '__all__'

class SubmissionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Submission
        fields = '__all__'
