ai_quizzer/requirements.txt
Django>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
google-generativeai>=0.1.1
psycopg2>=2.9  # Required if using PostgreSQL as the database